+++ runtime/datatypes/context.ts
@@ -129,7 +129,7 @@ export function $$mold(
 	
 	if(context.words.length > 0) {
 		buffer.push("\n");
-		const idt = "\t".repeat(indent);
+		const idt = " ".repeat(indent*4);
 
 		for(const word of context.words) {
 			const value = context.getWord(word);
@@ -141,7 +141,7 @@ export function $$mold(
 		}
 	}
 
-	buffer.push("\t".repeat(indent - 1));
+	buffer.push(" ".repeat((indent-1)*4));
 	buffer.push("]");
 	
 	return context.words.length > 0;