+++ runtime/datatypes/op.ts
@@ -1,4 +1,5 @@
 import * as Red from "../../red-types";
+import RedActions from "../actions";
 
 export function $$make(
 	_ctx:   Red.Context,
@@ -12,9 +13,55 @@ export function $$make(
 	}
 }
 
+export function $$mold(
+	ctx:    Red.Context,
+	op:     Red.Op,
+	buffer: string[],
+	indent: number,
+	_: RedActions.MoldOptions = {}
+): boolean {
+	const lastIndent = " ".repeat((indent-1)*4);
+	const thisIndent = " ".repeat(indent*4);
+	const value = op.func;
+
+	buffer.push("make op! [[");
+	
+	if(value.docSpec != null) {
+		buffer.push(RedActions.$$mold(ctx, value.docSpec).toJsString());
+	}
+
+	for(const arg of value.args) {
+		buffer.push("\n" + thisIndent);
+		RedActions.valueSendAction("$$mold", ctx, arg.name, buffer, indent + 1, _);
+		
+		if(arg.typeSpec != null) {
+			buffer.push(" ");
+			RedActions.valueSendAction("$$mold", ctx, arg.typeSpec, buffer, indent + 1, _);
+		}
+
+		if(arg.docSpec != null) {
+			buffer.push(" ");
+			RedActions.valueSendAction("$$mold", ctx, arg.docSpec, buffer, indent + 1, _);
+		}
+	}
+	
+	if(value.retSpec != null) {
+		buffer.push("\n" + thisIndent + "return: ");
+		RedActions.valueSendAction("$$mold", ctx, value.retSpec, buffer, indent + 1, _);
+	}
+
+	if(value.arity == 0 && value.refines.length == 0) {
+		buffer.push("]]");
+		return false;
+	} else {
+		buffer.push("\n" + lastIndent + "]]");
+		return true;
+	}
+}
+
 export function $$form(
 	_ctx:   Red.Context,
-	_value: Red.Op,
+	_op:    Red.Op,
 	buffer: string[],
 	_part?: number
 ): boolean {