+++ runtime/datatypes/percent.ts
@@ -1,29 +1,22 @@
 import * as Red from "../../red-types";
 import RedActions from "../actions";
 
-// $$make
-
-// $$to
-
 export function $$form(
 	_ctx:   Red.Context,
-	binary: Red.RawBinary,
+	value:  Red.RawPercent,
 	buffer: string[],
 	_part?: number
 ): boolean {
-	buffer.push("#{");
-	buffer.push(binary.bytes.toString("hex").toUpperCase());
-	buffer.push("}");
-	
+	buffer.push((value.value * 100).toString() + "%");
 	return false;
 }
 
 export function $$mold(
 	ctx:     Red.Context,
-	binary:  Red.RawBinary,
+	value:   Red.RawPercent,
 	buffer:  string[],
 	_indent: number,
 	_: RedActions.MoldOptions = {}
 ): boolean {
-	return $$form(ctx, binary, buffer, _.part);
+	return $$form(ctx, value, buffer, _.part);
 }